@@ -1,39 +1,34 @@
 ﻿using Dashboard.Client.Services.Interfaces;
 using Dashboard.Shared;
-using System.Collections.Generic;
-using System.Net.Http;
 using System.Net.Http.Json;
-using System.Threading.Tasks;
 
-namespace Dashboard.Client.Services
+namespace Dashboard.Client.Services;
+public class UserService : IUserService
 {
-    public class UserService : IUserService
+    private readonly HttpClient httpClient;
+
+    public UserService(HttpClient _httpClient)
     {
-        private readonly HttpClient httpClient;
+        httpClient = _httpClient;
+    }
 
-        public UserService(HttpClient _httpClient)
-        {
-            httpClient = _httpClient;
-        }
+    public async Task<List<UserModel>> GetUsers()
+    {
+        return await httpClient.GetFromJsonAsync<List<UserModel>>("api/users");
+    }
 
-        public async Task<List<UserModel>> GetUsers()
-        {
-            return await httpClient.GetFromJsonAsync<List<UserModel>>("api/users");
-        }
+    public async Task<List<RoleModel>> GetRoles()
+    {
+        return await httpClient.GetFromJsonAsync<List<RoleModel>>("api/users/roles");
+    }
 
-        public async Task<List<RoleModel>> GetRoles()
-        {
-            return await httpClient.GetFromJsonAsync<List<RoleModel>>("api/users/roles");
-        }
+    public async Task UpdateUser(UserModel user)
+    {
+        await httpClient.PostAsJsonAsync("api/users", user);
+    }
 
-        public async Task UpdateUser(UserModel user)
-        {
-            await httpClient.PostAsJsonAsync("api/users", user);
-        }
-
-        public async Task AddUser(UserModel user)
-        {
-            await httpClient.PutAsJsonAsync("api/users", user);
-        }
+    public async Task AddUser(UserModel user)
+    {
+        await httpClient.PutAsJsonAsync("api/users", user);
     }
 }
