@@ -1,86 +1,83 @@
-﻿using Dashboard.Infrastructure.Entity;
+﻿using Dashboard.Functions.Notifications.SendBug;
 using Dashboard.Server.Services.Interfaces;
-using Dashboard.Shared;
 using Dashboard.Shared.Identity;
+using MediatR;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Threading.Tasks;
 
-namespace Dashboard.Server.Controllers
+namespace Dashboard.Server.Controllers;
+
+[Route("auth")]
+[ApiController]
+public class IdentityController : ControllerBase
 {
-    [Route("auth")]
-    [ApiController]
-    public class IdentityController : ControllerBase
+    public record UserQuery(string username, string password);
+
+    private readonly IIdentityService identityService;
+    private readonly IMediator mediator;
+
+    public IdentityController(IMediator mediator, IIdentityService identityService)
     {
-        public record UserQuery(string username, string password);
+        this.identityService = identityService;
+        this.mediator = mediator;
+    }
 
-        private readonly IBugService _bugService;
-        private readonly IIdentityService _identityService;
 
-        public IdentityController(IBugService bugService, IIdentityService identityService)
+    [Route("login")]
+    [HttpPost]
+    [AllowAnonymous]
+    public async Task<IActionResult> Login([FromBody] UserQuery user)
+    {
+        try
         {
-            _bugService = bugService;
-            _identityService = identityService;
+            var retVal = await identityService.Login(user.username, user.password);
+            if (retVal.Succeeded)
+                return Ok(retVal.RetVal);
+            else
+                return NotFound(retVal.ErrorToString());
         }
-
-
-        [Route("login")]
-        [HttpPost]
-        [AllowAnonymous]
-        public async Task<IActionResult> Login([FromBody] UserQuery user)
+        catch (Exception ex)
         {
-            try
-            {
-                var retVal = await _identityService.Login(user.username, user.password);
-                if (retVal.Succeeded)
-                    return Ok(retVal.RetVal);
-                else
-                    return NotFound(retVal.ErrorToString());
-            }
-            catch (Exception ex)
-            {
-                await _bugService.SaveBug(ex.Message, "CORE API");
-                return StatusCode(500);
-            }
+            await mediator.Publish(new SendBugEvent(ex.Message, "CORE"));
+            return StatusCode(500);
         }
-        [Route("logout")]
-        [HttpPost]
-        //[Authorize]
-        public async Task<IActionResult> Logout()
+    }
+    [Route("logout")]
+    [HttpPost]
+    [Authorize]
+    public async Task<IActionResult> Logout()
+    {
+        try
         {
-            try
-            {
-                var accessToken = await HttpContext.GetTokenAsync("Bearer", "access_token");
-                await _identityService.Logout(HttpContext.User.Identity?.Name, accessToken);
-                return Ok();
-            }
-            catch (Exception ex)
-            {
-                await _bugService.SaveBug(ex.Message, "CORE API");
-                return StatusCode(500);
-            }
+            var accessToken = await HttpContext.GetTokenAsync("Bearer", "access_token");
+            await identityService.Logout(HttpContext.User.Identity?.Name, accessToken);
+            return Ok();
         }
-
-        [Route("refresh")]
-        [HttpPost]
-        [AllowAnonymous]
-        public async Task<IActionResult> Refresh([FromBody] RefreshTokenRequest request)
+        catch (Exception ex)
         {
-            try
-            {
-                var retVal = await _identityService.RefreshToken(request);
-                if (retVal.Succeeded)
-                    return Ok(retVal.RetVal);
-                else
-                    return BadRequest(retVal.ErrorToString());
-            }
-            catch (Exception ex)
-            {
-                await _bugService.SaveBug(ex.Message, "CORE API");
-                return StatusCode(500);
-            }
+            await mediator.Publish(new SendBugEvent(ex.Message, "CORE"));
+            return StatusCode(500);
+        }
+    }
+
+    [Route("refresh")]
+    [HttpPost]
+    [AllowAnonymous]
+    public async Task<IActionResult> Refresh([FromBody] RefreshTokenRequest request)
+    {
+        try
+        {
+            var retVal = await identityService.RefreshToken(request);
+            if (retVal.Succeeded)
+                return Ok(retVal.RetVal);
+            else
+                return BadRequest(retVal.ErrorToString());
+        }
+        catch (Exception ex)
+        {
+            await mediator.Publish(new SendBugEvent(ex.Message, "CORE"));
+            return StatusCode(500);
         }
     }
 }
