@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Dashboard.Shared
+﻿namespace Dashboard.Shared;
+public class BugModel
 {
-    public class BugModel
-    {
-        public int Id { get; set; }
-        public string Message { get; set; }
-        public string System { get; set; }
-        public DateTime Date { get; set; }
-    }
+    public int Id { get; set; }
+    public string Message { get; set; }
+    public string System { get; set; }
+    public DateTime Date { get; set; }
 }
