@@ -1,8 +1,6 @@
-﻿namespace Dashboard.Shared
+﻿namespace Dashboard.Shared;
+public class ConfigurationModel
 {
-    public class ConfigurationModel
-    {
-        public int WatcherWorkerDelayTime { get; set; }
-        public int TokenExpirationTime { get; set; }
-    }
+    public int WatcherWorkerDelayTime { get; set; }
+    public int TokenExpirationTime { get; set; }
 }
