@@ -1,8 +1,6 @@
-﻿namespace Dashboard.Shared.Identity
+﻿namespace Dashboard.Shared.Identity;
+public class RefreshTokenRequest
 {
-    public class RefreshTokenRequest
-    {
-        public string AccessToken { get; set; }
-        public string RefreshToken { get; set; }
-    }
+    public string AccessToken { get; set; }
+    public string RefreshToken { get; set; }
 }
