@@ -1,8 +1,6 @@
-﻿namespace Dashboard.Shared
+﻿namespace Dashboard.Shared;
+public class RoleModel
 {
-    public class RoleModel
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
+    public int Id { get; set; }
+    public string Name { get; set; }
 }
