@@ -1,11 +1,7 @@
-﻿using Dashboard.Client.Services;
-using Dashboard.Client.Services.Interfaces;
-using Dashboard.Shared;
+﻿using Dashboard.Client.Services.Interfaces;
 using Dashboard.Shared.Identity;
 using Microsoft.AspNetCore.Components.Authorization;
-using System;
 using System.Security.Claims;
-using System.Threading.Tasks;
 
 namespace Dashboard.Client.Authentication
 {
