@@ -1,10 +1,6 @@
-﻿using Dashboard.Client.Services;
-using Dashboard.Shared;
+﻿using Dashboard.Shared;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web.Virtualization;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using System.Linq;
 using Dashboard.Client.Services.Interfaces;
 
 namespace Dashboard.Client.Pages
