@@ -1,10 +1,7 @@
-﻿using Dashboard.Client.Services;
-using Dashboard.Client.Services.Interfaces;
+﻿using Dashboard.Client.Services.Interfaces;
 using Dashboard.Shared;
 using Microsoft.AspNetCore.Components;
-using System;
 using System.ComponentModel.DataAnnotations;
-using System.Threading.Tasks;
 
 namespace Dashboard.Client.Pages
 {
